% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{addExport}
\alias{addExport}
\title{Add libraries dependencies used in export \link{visExport}}
\usage{
addExport(graph, pdf = TRUE)
}
\arguments{
\item{graph}{: a visNetwork object}

\item{pdf}{: boolean. Add jsPDF or not ?}
}
\value{
\code{graph} htmlwidget with dependencies attached.
}
\description{
Add libraries dependencies used in export \link{visExport}
}
