% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dependencies.R
\name{addFontAwesome}
\alias{addFontAwesome}
\title{Use fontAwesome icons in visNetwork \code{graph}}
\usage{
addFontAwesome(graph)
}
\arguments{
\item{graph}{: a visNetwork object}
}
\value{
\code{graph} htmlwidget with Font-Awesome dependencies attached.
}
\description{
Add \href{http://fortawesome.github.io/Font-Awesome/}{Font-Awesome} for styling
  our \code{graph} with beautiful, professional icons.  Please note
  that you'll already have these icons if using Shiny.
}
\examples{
#╦ use fontAwesome icons using groups or nodes options
# font-awesome is not part of dependencies. use addFontAwesome() if needed.
# http://fortawesome.github.io/Font-Awesome/

nodes <- data.frame(id = 1:3, group = c("B", "A", "B"))
edges <- data.frame(from = c(1,2), to = c(2,3))

visNetwork(nodes, edges) \%>\%
  visGroups(groupname = "A", shape = "icon", icon = list(code = "f0c0", size = 75)) \%>\%
  visGroups(groupname = "B", shape = "icon", icon = list(code = "f007", color = "red")) \%>\%
  addFontAwesome()

nodes <- data.frame(id = 1:3)
edges <- data.frame(from = c(1,2), to = c(1,3))

visNetwork(nodes, edges) \%>\%
  visNodes(shape = "icon", icon = list( face ='FontAwesome', code = "f0c0")) \%>\%
  addFontAwesome()
}

