% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visClustering.R
\name{visClusteringByConnection}
\alias{visClusteringByConnection}
\title{Network visualization clustering options - by node id}
\usage{
visClusteringByConnection(graph, nodes)
}
\arguments{
\item{graph}{: a visNetwork object}

\item{nodes}{: Character/vector. id of nodes we want to cluster}
}
\description{
Network visualization clustering options - by node id
}
\examples{

set.seed(124)
nodes <- data.frame(id = 1:10, color = c(rep("blue", 6), rep("red", 3), rep("green", 1)))
edges <- data.frame(from = round(runif(6)*10), to = round(runif(6)*10))

 visNetwork(nodes, edges) \%>\%
   visClusteringByConnection(nodes = 9)
     
}
