% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/visClustering.R
\name{visClusteringOutliers}
\alias{visClusteringOutliers}
\title{Network visualization clustering options}
\usage{
visClusteringOutliers(graph, clusterFactor = 0.9, stabilize = FALSE)
}
\description{
Network visualization clustering options.
}
\examples{
set.seed(123)
nodes <- data.frame(id = 1:100)
edges <- data.frame(from = round(runif(60)*100), to = round(runif(60)*100))

visNetwork(nodes, edges) \%>\%
 visClusteringOutliers(clusterFactor = 1) \%>\%
 visLayout(randomSeed = 123)
}

