% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visExport.R
\name{visExport}
\alias{visExport}
\title{Network export configuration}
\usage{
visExport(
  graph,
  type = "png",
  name = "network",
  label = paste0("Export as ", type),
  background = "#fff",
  float = "right",
  style = NULL,
  loadDependencies = TRUE,
  ...
)
}
\arguments{
\item{graph}{: a visNetwork object}

\item{type}{: Type of export. One of "png" (default), "jpeg" or "pdf"}

\item{name}{: name of imgage, default to "network"}

\item{label}{: Label on button, default to "Export as png/jpeg/pdf"}

\item{background}{: background color, default to white (#fff). Work only if network background is transparent.}

\item{float}{: button position, default to "right"}

\item{style}{: button css style.}

\item{loadDependencies}{/ Boolean. TRUE by default. Load libraries for export (fileSaver, Blob, canvastoBlob, html2canvas, jsPDF)}

\item{...}{: arguments for  \link[visNetwork]{addExport}}
}
\description{
Network export configuration. This function only work within shiny or a web browser.
}
\examples{

\dontrun{

nodes <- data.frame(id = 1:3, group = c("B", "A", "B"))
edges <- data.frame(from = c(1,2), to = c(2,3))

visNetwork(nodes, edges) \%>\%
 visGroups(groupname = "A", color = "red") \%>\%
 visGroups(groupname = "B", color = "lightblue") \%>\%
 visLegend() \%>\% visExport() 
 
visNetwork(nodes, edges) \%>\%
 visGroups(groupname = "A", color = "red") \%>\%
 visGroups(groupname = "B", color = "lightblue") \%>\%
 visLegend() \%>\% visExport(type = "jpeg", name = "export-network", 
   float = "left", label = "Save network", background = "purple", style= "") 
 
}
}
\references{
See online documentation \url{https://datastorm-open.github.io/visNetwork/}
}
\seealso{
\link[visNetwork]{visSave}
}
