% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visGetUtils.R
\name{visGetBoundingBox}
\alias{visGetBoundingBox}
\title{Method getBoundingBox, with shiny only.}
\usage{
visGetBoundingBox(graph, id, input = paste0(graph$id, "_boundingBox"))
}
\arguments{
\item{graph}{: a \code{\link{visNetworkProxy}}  object}

\item{id}{: a node or edge id}

\item{input}{: name of shiny input created. Default to paste0(graph$id, "_boundingBox")}
}
\description{
Method getBoundingBox, with shiny only. Returns a bounding box for the node including label in the format. These values are in canvas space.
}
\examples{
\dontrun{

# have a look to : 
shiny::runApp(system.file("shiny", package = "visNetwork"))

}

}
\references{
See online documentation \url{http://datastorm-open.github.io/visNetwork/}
}
\seealso{
\link{visNodes} for nodes options, \link{visEdges} for edges options, \link{visGroups} for groups options, 
\link{visLegend} for adding legend, \link{visOptions} for custom option, \link{visLayout} & \link{visHierarchicalLayout} for layout, 
\link{visPhysics} for control physics, \link{visInteraction} for interaction, \link{visNetworkProxy} & \link{visFocus} & \link{visFit} for animation within shiny,
\link{visDocumentation}, \link{visEvents}, \link{visConfigure} ...
}
