% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visLegend.R
\name{visLegend}
\alias{visLegend}
\title{Add a legend on a visNetwork object}
\usage{
visLegend(graph, enabled = TRUE, width = 0.2, nodes = NULL,
  edges = NULL)
}
\arguments{
\item{graph}{: a visNetwork object}

\item{dragNodes}{: Boolean. Default to true. When true, the nodes that are not fixed can be dragged by the user.}
}
\description{
Add a legend on a visNetwork object
}
\examples{
# minimal example
nodes <- data.frame(id = 1:3, group = c("B", "A", "B"))
edges <- data.frame(from = c(1,2), to = c(2,3))

# default, on group
visNetwork(nodes, edges) \%>\%
  visGroups(groupname = "A", color = "blue") \%>\%
  visGroups(groupname = "B", color = "yellow") \%>\%
  visLegend()

# default, on group, control width
visNetwork(nodes, edges) \%>\%
 visGroups(groupname = "A", color = "blue") \%>\%
  visGroups(groupname = "B", color = "yellow") \%>\%
  visLegend(width = 0.2)

# passing custom nodes and/or edges
nodesleg <- data.frame(label = c("Group A", "Group B"), shape = c( "ellipse"), color = c("yellow", "blue"),
 title = "Informations")

visNetwork(nodes, edges) \%>\%
  visGroups(groupname = "A", color = "blue") \%>\%
  visGroups(groupname = "B", color = "yellow") \%>\%
  visLegend(nodes = nodesleg)

edgesled <- data.frame(color = c("blue", "yellow"), label = c("A", "B"), arrows =c("to", "from"))

visNetwork(nodes, edges) \%>\%
  visGroups(groupname = "A", color = "blue") \%>\%
  visGroups(groupname = "B", color = "yellow") \%>\%
  visLegend(edges = edgesled)

visNetwork(nodes, edges) \%>\%
  visGroups(groupname = "A", color = "blue") \%>\%
  visGroups(groupname = "B", color = "yellow") \%>\%
  visLegend(edges = edgesled, nodes = nodesleg)
}
\seealso{
\link{visOptions}, \link{visNodes}, \link{visEdges}, \link{visGroups}, \link{visEvents}
}

