% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/visNodes.R
\name{visNodes}
\alias{visNodes}
\title{Network visualization nodes options}
\usage{
visNodes(graph, borderWidth = NULL, borderWidthSelected = NULL,
  customScalingFunction = NULL, color = NULL, fontColor = NULL,
  fontFace = NULL, fontSize = NULL, scaleFontWithValue = NULL,
  fontSizeMin = NULL, fontSizeMax = NULL, fontSizeMaxVisible = NULL,
  fontDrawThreshold = NULL, fontFill = NULL, fontStrokeWidth = NULL,
  fontStrokeColor = NULL, shape = NULL, image = NULL,
  brokenImage = NULL, mass = NULL, widthMin = NULL, widthMax = NULL,
  radius = NULL, radiusMin = NULL, radiusMax = NULL,
  iconFontFace = NULL, icon = NULL, iconSize = NULL)
}
\arguments{
\item{borderWidth}{: Number. Default to 1. The width of the border of the node when it is not selected, automatically limited by the width of the node.}

\item{borderWidthSelected}{: Number. Undefined. The width of the border of the node when it is selected. If left at undefined, double the borderWidth will be used.}

\item{customScalingFunction}{: Function. This is a function you can override to make the nodes scale the way you want them based on their values.}

\item{color}{: String | named list.    Color for the node. Can be just one color, or a list with several elements :
\itemize{
 \item{"background"}{ : String. Default to '#97C2FC'. Background color for the node.}
 \item{"border"}{ : String. Default to '#2B7CE9'. Border color for the node.}
 \item{"highlight"}{ : String | named list,     Color of the node when selected.
   \itemize{
     \item{"background"}{ : String. Default to '#97C2FC'. Background color for the node when selected.}
     \item{"border"}{ : String. Default to '#2B7CE9'. Border color for the node when selected.}
   }
 }
 \item{"hover"}{ : named list, when the hover option is enabled
   \itemize{
     \item{"background"}{ : String. Default to '#2B7CE9'. Border color of the node when selected.}
     \item{"border"}{ : String. Default to '#2B7CE9'. Border color of the node when the node is hovered over and the hover option is enabled.}
   }
 }
}}

\item{fontColor}{: String. Default to 'black'. Font color for label in the node.}

\item{fontFace}{: String. Default to 'verdana'. Font face for label in the node, for example "verdana" or "arial".}

\item{fontSize}{: Number. Default to 14. Font size in pixels for label in the node.}

\item{scaleFontWithValue}{: Boolean. Default to false. When using values, you can let the font scale with the size of the nodes if you enable the this option.}

\item{fontSizeMin}{: Number. Default to 14. When using values, you can let the font scale with the size of the nodes if you enable the scaleFontWithValue option. This is the minimum value of the fontSize.}

\item{fontSizeMax}{: Number. Default to 30. When using values, you can let the font scale with the size of the nodes if you enable the scaleFontWithValue option. This is the maximum value of the fontSize.}

\item{fontSizeMaxVisible}{: Number. Default to 30. When using values, you can let the font scale with the size of the nodes if you enable the scaleFontWithValue option. If you have a wide distribution of values and have a large max fontSize, the text will become huge if you zoom in on it. This option limits the percieved fontSize to avoid this. If you set it to 20, no label will be larger than fontsize 20 (at scale = 1) regardless of the scale.}

\item{fontDrawThreshold}{: Number. Default to 3. When zooming out, the text becomes smaller. This option sets the minimum size of the label before not being drawn. Just like the fontSizeMaxVisible option, this is the relative fontSize (fontSize * scale). You can combine this with the min and max values to have the labels of influential nodes show earlier when zooming in.}

\item{fontFill}{: String. Undefined. If a color is supplied, there will be a background color behind the label. If left undefined, no background color is shown.}

\item{fontStrokeWidth}{: Number. Default to 0. The width of the label stroke (border around label's text) in pixels.}

\item{fontStrokeColor}{: String. Default to 'white'. The color of the label stroke.}

\item{shape}{: String. Default to 'ellipse'. Define the shape for the node. Choose from ellipse (default), circle, box, database, image, circularImage, label, dot, star, triangle, triangleDown, square and icon. The shapes dot, star, triangle, triangleDown, and square, are scalable. The size is determined by the properties radius or value.}

\item{image}{: String. Undefined. Default image url for the nodes. only applicable to shape image.}

\item{brokenImage}{: String. Undefined. Image url to use in the event that the url specified in the image property fails to load. only applicable to shape image.}

\item{mass}{: Number. Default to 1. When using the Barnes Hut simulation method (which is selected by default), the mass of a node determines the gravitational repulsion during the simulation. Higher mass will push other nodes further away. Preferably use the physics configuration to alter the simulation.}

\item{widthMin}{: Number. Default to 16. The minimum width for a scaled image. Only applicable to shape image. This only does something if you supply a value.}

\item{widthMax}{: Number. Default to 64. The maximum width for a scaled image. Only applicable to shape image. This only does something if you supply a value.}

\item{radius}{: Number. Default to 10. The default radius for a node. Only applicable to shapes dot, star, triangle, triangleDown, and square.}

\item{radiusMin}{: Number. Default to 10. The minimum radius for a scaled node. Only applicable to shapes dot, star, triangle, triangleDown, and square. This only does something if you supply a value.}

\item{radiusMax}{: Number. Default to 30. The maximum radius for a scaled node. Only applicable to shapes dot, star, triangle, triangleDown, and square. This only does something if you supply a value.}

\item{iconFontFace}{: String. Undefined. Font face for icons, for example FontAwesome or Ionicon. You have to link to the css defining the font by yourself (see Examples)}

\item{icon}{: String. Undefined. Unicode of the icon f.e. \uf0c0 (user-icon in FontAwesome)}

\item{iconSize}{: Number. Default to 50. Size of the icon}
}
\description{
Network visualization nodes options \url{http://visjs.org/docs/network.html#Nodes_configuration}
}

