% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visRpart.R
\name{visRpart}
\alias{visRpart}
\title{Plotting rpart objects}
\usage{
visRpart(object, main = "", direction = "UD", fontSize = 11,
  colorVar = NULL, colorMod = NULL, colorEdges = "#8181F7",
  legend = TRUE, legendWidth = 0.1, legendNcol = 1,
  highlightNearest = list(enabled = TRUE, degree = list(from = 50000, to = 0),
  hover = TRUE, algorithm = "hierarchical"), collapse = list(enabled = TRUE,
  fit = TRUE, unselect = TRUE, clusterOptions = list(fixed = TRUE, physics =
  FALSE)), tooltipDelay = 500, rules = TRUE, digits = 3,
  height = "500px", width = "100\%")
}
\arguments{
\item{object}{\code{rpart}, rpart object}

\item{main}{\code{character}, Graph title}

\item{direction}{\code{character}, The direction of the hierarchical layout.
The available options are: UD, DU, LR, RL. To simplify:
up-down, down-up, left-right, right-left. Default UD}

\item{fontSize}{\code{numeric}, size of label. Defaut to 11}

\item{colorVar}{\code{data.frame} 2 columns :
'variable' with names of variables X
'color' with colors (in hexa)}

\item{colorMod}{\code{data.frame} 2 columns :
'modality' with levels of Y
'color' with colors (in hexa)}

\item{colorEdges}{\code{character} color of edges, in hexa. Default #8181F7}

\item{legend}{\code{boolean}, add legend ? Default TRUE}

\item{legendWidth}{\code{numeric}, legend width, between 0 and 1. Default 0.1}

\item{legendNcol}{\code{numeric}, number of column for legend. Default 1}

\item{highlightNearest}{\code{boolean}, Highlight nearest nodes. See \link{visOptions}}

\item{collapse}{\code{boolean}, collapse or not using double click on a node ? See \link{visOptions}}

\item{tooltipDelay}{\code{numeric}, delay before tooltips 
apparition in millisecond. Default 500}

\item{rules}{\code{boolean}, add rules in tooltips ? Default TRUE}

\item{digits}{\code{numeric}, number of digits. Defaut to 3}

\item{height}{\code{character}, defaut to "500px"}

\item{width}{\code{character}, defaut to "100%"}
}
\description{
Function to plot rpart objects
}
\examples{

\dontrun{
library(rpart)

# Basic classification tree
res <- rpart(Species~., data=iris)
visRpart(res)
visRpart(res, direction = 'LR', main = "Iris classification Tree", fontSize = 15)

# Basic regression tree
res <- rpart(Petal.Length~., data=iris)
visRpart(res)

# disable rules in tooltip, and render faster
visRpart(res, rules = FALSE, tooltipDelay = 0)

# Complex tree
data("solder")
res<- rpart(Opening~., data=solder, control = (rpart.control(cp = 0.000005)))
visRpart(res, height = "800px", fontSize = 15)

# Change color
colorVar <- data.frame(variable = names(solder), color = c("#339933", "#b30000","#4747d1",
                                                           "#88cc00", "#9900ff","#247856"))

colorMod <- data.frame(modality = unique(solder$Opening), color = 
                              c("#AA00AA", "#CDAD15", "#213478"))

visRpart(res, colorEdges = "#000099", colorVar = colorVar, 
        colorMod = colorMod)

}

}

