% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visSelectNodes.R
\name{visSelectNodes}
\alias{visSelectNodes}
\title{Function to select node(s) from network, with shiny only.}
\usage{
visSelectNodes(graph, id, highlightEdges = TRUE, clickEvent = TRUE)
}
\arguments{
\item{graph}{: a \code{\link[pkg:visNetwork]{visNetworkProxy}}  object}

\item{id}{: vector of id, node(s) to select}

\item{highlightEdges}{: Boolean. highlight Edges also ? Default to TRUE}

\item{clickEvent}{: Boolean. Launch click event ? (highlightNearest for example) Default to TRUE}
}
\description{
Function to select node(s) from network, with shiny only.
}
\examples{
\dontrun{

# have a look to : 
shiny::runApp(system.file("shiny", package = "visNetwork"))

}

}
\references{
See online documentation \url{https://datastorm-open.github.io/visNetwork/}
}
\seealso{
\link[pkg:visNetwork]{visNodes} for nodes options, \link[pkg:visNetwork]{visEdges} for edges options, \link[pkg:visNetwork]{visGroups} for groups options, 
\link[pkg:visNetwork]{visLegend} for adding legend, \link[pkg:visNetwork]{visOptions} for custom option, \link[pkg:visNetwork]{visLayout} & \link[pkg:visNetwork]{visHierarchicalLayout} for layout, 
\link[pkg:visNetwork]{visPhysics} for control physics, \link[pkg:visNetwork]{visInteraction} for interaction, \link[pkg:visNetwork]{visNetworkProxy} & \link[pkg:visNetwork]{visFocus} & \link[pkg:visNetwork]{visFit} for animation within shiny,
\link[pkg:visNetwork]{visDocumentation}, \link[pkg:visNetwork]{visEvents}, \link[pkg:visNetwork]{visConfigure} ...
}
