% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visSetData.R
\name{visSetData}
\alias{visSetData}
\title{Network visualization setData method}
\usage{
visSetData(graph, nodes = NULL, edges = NULL)
}
\arguments{
\item{graph}{: a \code{\link[pkg:visNetwork]{visNetworkProxy}}  object}

\item{nodes}{: data.frame with nodes informations. Needed at least column "id". See \link[pkg:visNetwork]{visNodes}}

\item{edges}{: data.frame with edges informations. Needed at least columns "from" and "to". See \link[pkg:visNetwork]{visEdges}}
}
\description{
For use setData() method in a shiny app. For full documentation, have a look at \link[pkg:visNetwork]{visDocumentation}.
}
\examples{
\dontrun{

# have a look to : 
shiny::runApp(system.file("shiny", package = "visNetwork"))

}

}
\references{
See online documentation \url{https://datastorm-open.github.io/visNetwork/}
}
\seealso{
\link[pkg:visNetwork]{visNodes} for nodes options, \link[pkg:visNetwork]{visEdges} for edges options, \link[pkg:visNetwork]{visGroups} for groups options, 
\link[pkg:visNetwork]{visLegend} for adding legend, \link[pkg:visNetwork]{visOptions} for custom option, \link[pkg:visNetwork]{visLayout} & \link[pkg:visNetwork]{visHierarchicalLayout} for layout, 
\link[pkg:visNetwork]{visPhysics} for control physics, \link[pkg:visNetwork]{visInteraction} for interaction, \link[pkg:visNetwork]{visNetworkProxy} & \link[pkg:visNetwork]{visFocus} & \link[pkg:visNetwork]{visFit} for animation within shiny,
\link[pkg:visNetwork]{visDocumentation}, \link[pkg:visNetwork]{visEvents}, \link[pkg:visNetwork]{visConfigure} ...
}
