% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visSetTitle.R
\name{visSetTitle}
\alias{visSetTitle}
\title{Set title, subtitle, and footer using \code{visNetworkProxy}}
\usage{
visSetTitle(graph, main = NULL, submain = NULL, footer = NULL)
}
\arguments{
\item{graph}{: a \code{\link[pkg:visNetwork]{visNetworkProxy}}  object}

\item{main}{: For add a title. Character or a named list.
\itemize{
 \item "text" Character. Title.
 \item "style" Optional. Character. HTML style of title.
 \item 'hidden' Optional. Boolean. Force title to be hidden 
}}

\item{submain}{: For add a subtitle. Character or a named list.
\itemize{
 \item "text" Character. Subtitle.
 \item "style" Optional. Character. HTML style of submain.
 \item 'hidden' Optional. Boolean. Force submain to be hidden 
}}

\item{footer}{: For add a footer. Character or a named list.
\itemize{
 \item "text" Character. footer.
 \item "style" Optional. Character. HTML style of footer.
 \item 'hidden' Optional. Boolean. Force footer to be be hidden 
}}
}
\description{
Set title, subtitle, and footer using \code{visNetworkProxy}
}
\examples{
\dontrun{

# have a look to : 
shiny::runApp(system.file("shiny", package = "visNetwork"))

}

}
\references{
See online documentation \url{https://datastorm-open.github.io/visNetwork/}
}
\seealso{
\link[pkg:visNetwork]{visNodes} for nodes options, \link[pkg:visNetwork]{visEdges} for edges options, \link[pkg:visNetwork]{visGroups} for groups options, 
\link[pkg:visNetwork]{visLegend} for adding legend, \link[pkg:visNetwork]{visOptions} for custom option, \link[pkg:visNetwork]{visLayout} & \link[pkg:visNetwork]{visHierarchicalLayout} for layout, 
\link[pkg:visNetwork]{visPhysics} for control physics, \link[pkg:visNetwork]{visInteraction} for interaction, \link[pkg:visNetwork]{visNetworkProxy} & \link[pkg:visNetwork]{visFocus} & \link[pkg:visNetwork]{visFit} for animation within shiny,
\link[pkg:visNetwork]{visDocumentation}, \link[pkg:visNetwork]{visEvents}, \link[pkg:visNetwork]{visConfigure} ...
}
