% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/visTooltips.R
\name{visTooltips}
\alias{visTooltips}
\title{Network visualization tooltip options}
\usage{
visTooltips(graph, delay = 300, fontColor = "black", fontSize = 14,
  fontFace = "verdana", color.background = "#FFFFFF",
  color.border = "#666")
}
\arguments{
\item{delay}{: Number. Default to 300. Time in milliseconds a user must hover over a node or edge before a tooltip appears.}

\item{fontColor}{: String. Default to "black". Default color for tooltip text.}

\item{fontSize}{: Number. Default to 14. Size in pixels of tooltip text.}

\item{fontFace}{: String. Default to "verdana". Font family to used for tooltip text.}

\item{color.background}{: String. Default to "#FFFFFF". Background color for the node.}

\item{color.border}{: String. Default to "#666". Border color for the node.}
}
\description{
Network visualization tooltip options. See \url{http://visjs.org/docs/network.html#Tooltips}.
}
\examples{
nodes <- data.frame(id = 1:3, title = paste0("<p>", 1:3,"<br> tooltip</p>"))
edges <- data.frame(from = c(1,2), to = c(1,3))

visNetwork(nodes, edges) \%>\%
 visTooltips(delay = 100, color.background = "red", fontColor = "white")
}

