% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visClustering.R
\name{visClusteringOutliers}
\alias{visClusteringOutliers}
\title{Network visualization clustering options - outliers}
\usage{
visClusteringOutliers(graph, clusterFactor = 0.9, stabilize = FALSE)
}
\arguments{
\item{graph}{: a visNetwork object}

\item{clusterFactor}{: Number, from 0 to 1. 0.9 by default}

\item{stabilize}{: Boolean, default to false}
}
\description{
Network visualization clustering options - outliers
}
\examples{
 
nodes <- data.frame(id = 1:10)
edges <- data.frame(from = c(1,1,10,2,6,7,8,9,10), 
                    to = c(2,3,4,5,2,5,6,7,9))

visNetwork(nodes, edges) \%>\%
 visClusteringOutliers(1)

 
}
