% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visConfigure.R
\name{visConfigure}
\alias{visConfigure}
\title{Network visualization configure options}
\usage{
visConfigure(graph, enabled = NULL, filter = NULL, container = NULL,
  showButton = NULL)
}
\arguments{
\item{graph}{: a visNetwork object}

\item{enabled}{: Boolean. Default to true. Toggle the configuration interface on or off. This is an optional parameter. If left undefined and any of the other properties of this object are defined, this will be set to true.}

\item{filter}{: String, Array, Boolean, Function. Default to true. When a boolean, true gives you all options, false will not show any. If a string is supplied, any combination of the following is allowed: nodes, edges, layout, interaction, manipulation, physics, selection, renderer. Feel free to come up with a fun seperating character. Finally, when supplied an array of strings, any of the previously mentioned fields are accepted.}

\item{container}{: DOM element. This allows you to put the configure list in another HTML container than below the network.}

\item{showButton}{: Boolean. Default to true. Show the generate options button at the bottom of the configurator.}
}
\description{
Network visualization configure options. For full documentation, have a look at \link{visDocumentation}.
}
\examples{

\\donrun{

nodes <- data.frame(id = 1:3, title = paste0("<p>", 1:3,"<br> tooltip</p>"))
edges <- data.frame(from = c(1,2), to = c(1,3))

# don't look in RStudio viewer
visNetwork(nodes, edges) \%>\%
 visConfigure(enabled = TRUE, filter = "interaction")
 
# shiny
# Define UI
ui <- shinyUI(fluidPage(
  fluidRow(column(4, div(id = "config")), 
           column(8,visNetworkOutput("network"))
  )
))

# Define SERVER
server <- shinyServer(function(input, output) {
  
  output$network <- renderVisNetwork({
    nodes <- data.frame(id = 1:15, label = paste("Label", 1:15),
                        group = sample(LETTERS[1:3], 15, replace = TRUE))
    
    edges <- data.frame(from = trunc(runif(15)*(15-1))+1,
                        to = trunc(runif(15)*(15-1))+1)
    
    visNetwork(nodes, edges) \%>\% visConfigure(enabled = TRUE, container = "config")
  })
})

# Run the application 
shinyApp(ui = ui, server = server)  
 
}

}
\references{
See online documentation \url{http://datastorm-open.github.io/visNetwork/}
}
\seealso{
\link{visNodes} for nodes options, \link{visEdges} for edges options, \link{visGroups} for groups options, 
\link{visLegend} for adding legend, \link{visOptions} for custom option, \link{visLayout} & \link{visHierarchicalLayout} for layout, 
\link{visPhysics} for control physics, \link{visInteraction} for interaction, \link{visNetworkProxy} & \link{visFocus} & \link{visFit} for animation within shiny,
\link{visDocumentation}, \link{visEvents}, \link{visConfigure} ...
}
