% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/visDocumentation.R
\name{visDocumentation}
\alias{visDocumentation}
\title{View full documentation of vis.js on network}
\usage{
visDocumentation()
}
\description{
View full documentation of vis.js on network
}
\examples{
# minimal example
visDocumentation()
}
\seealso{
\link{visOptions}, \link{visNodes}, \link{visEdges}, \link{visGroups}, \link{visEvents}
}

