% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visDocumentation.R
\name{visDocumentation}
\alias{visDocumentation}
\title{View full documentation of vis.js on network}
\usage{
visDocumentation(viewer = getOption("viewer"))
}
\arguments{
\item{viewer}{: Set to NULL to open in a browser}
}
\description{
View full documentation of vis.js on network
}
\examples{

# minimal example
\dontrun{
visDocumentation()
visDocumentation(NULL)
}

}
\references{
See online documentation \url{http://datastorm-open.github.io/visNetwork/}
}
\seealso{
\link{visNodes} for nodes options, \link{visEdges} for edges options, \link{visGroups} for groups options, 
\link{visLegend} for adding legend, \link{visOptions} for custom option, \link{visLayout} & \link{visHierarchicalLayout} for layout, 
\link{visPhysics} for control physics, \link{visInteraction} for interaction, \link{visNetworkProxy} & \link{visFocus} & \link{visFit} for animation within shiny,
\link{visDocumentation}, \link{visEvents}, \link{visConfigure} ...
}
