% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/visNetworkEvents.R
\name{visEvents}
\alias{visEvents}
\title{Network visualization events}
\usage{
visEvents(graph, animationFinished = NULL, select = NULL, click = NULL,
  doubleClick = NULL, hoverNode = NULL, blurNode = NULL, resize = NULL,
  dragStart = NULL, dragEnd = NULL, startStabilization = NULL,
  stabilizationIterationsDone = NULL, stabilized = NULL,
  viewChanged = NULL, zoom = NULL)
}
\arguments{
\item{graph}{: a visNetwork object}

\item{animationFinished}{: Fired after an animation is finished}

\item{select}{: Fired after the user selects or deselects a node by clicking it. Not fired when the method setSelectionis executed.}

\item{click}{: Fired after the user clicks or taps on a touchscreen.}

\item{doubleClick}{: Fired after the user double clicks or double taps on a touchscreen.}

\item{hoverNode}{: Fired when the mouse is moved over a node (assuming the hover option is enabled).}

\item{blurNode}{: Fired when the mouse is moved off a node (assuming the hover option is enabled).}

\item{resize}{: Fired when the size of the canvas has been resized, either by a redraw call when the container div has changed in size, a setSize() call with new values or a setOptions() with new width and/or height values.}

\item{dragStart}{: Fired when a node is being dragged.}

\item{dragEnd}{: Fired when the dragging of a node(s) has ended.}

\item{startStabilization}{: Fired once when the network starts the physics calculation. This ends with the stabilized event.     none}

\item{stabilizationIterationsDone}{: Fired once when the network finished the initial stabilization run. This is fired REGARDLESS if the network has stabilized. It only means that the amount of configured stabilizationIterations have been completed.     none}

\item{stabilized}{: Fired every time the network has been stabilized. This event can be used to trigger the .storePositions() function after stabilization. Fired with an object having the following properties:}

\item{viewChanged}{: Fired when the view has changed. This is when the network has moved or zoomed.     none}

\item{zoom}{: Fired when the network has zoomed. This event can be used to trigger the .storePositions() function after stabilization.}
}
\description{
Network visualization events.
}
\examples{
nodes <- data.frame(id = 1:3)
edges <- data.frame(from = c(1,2), to = c(1,3))

visNetwork(nodes, edges) \%>\%
 visEvents(select = "function onSelect (properties) {
     alert('selected nodes: ' + properties.nodes);}")
}

