% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visHclust.R
\name{visHclust}
\alias{visHclust}
\title{Visualize Hierarchical cluster analysis.}
\usage{
visHclust(data, main = "", submain = "", footer = "",
  distColumns = NULL, distMethod = "euclidean", hclustMethod = "complete",
  cutree = 0, tooltipColumns = 1:ncol(data), colorEdges = "black",
  colorGroups = substr(rainbow(cutree), 1, 7), nodesSize = 30,
  highlightNearest = TRUE, height = "600px", width = "100\%")
}
\arguments{
\item{data}{\code{data.frame} data.}

\item{main}{For add a title. See \link{visNetwork}}

\item{submain}{For add a subtitle. See \link{visNetwork}}

\item{footer}{For add a footer. See \link{visNetwork}}

\item{distColumns}{\code{numeric} indice of columns used for compute distance. 
If NULL (default), keep all \code{numeric} and \code{integer} columns. 
If Not NULL, we keep only  \code{numeric} and \code{integer} columns}

\item{distMethod}{\code{character} the distance measure to be used for dist function. Default to 'euclidean'. See \code{\link[stats]{dist}}.}

\item{hclustMethod}{\code{character} the agglomeration method to be used for hclust function. Default to 'complete'. See \code{\link[stats]{hclust}}.}

\item{cutree}{\code{numeric} or \code{integer} desired number of groups. Default to 0}

\item{tooltipColumns}{\code{numeric} indice of columns used in tooltip. All by default.
So, we add a boxplot or a pie focus on sub-population and all population using \code{sparkline} package.}

\item{colorEdges}{\code{character} color of edges. Default to 'black'}

\item{colorGroups}{\code{character}, color for group in exa ("#00FF00"). Default rainbow.}

\item{nodesSize}{\code{numeric} size for nodes. Default to 30}

\item{highlightNearest}{\code{boolean} highlight sub-tree on click.}

\item{height}{\code{character}, default to "600px"}

\item{width}{\code{character}, default to "100\%"}
}
\description{
Visualize Hierarchical cluster analysis \code{hclust}. This function compute distance using \code{dist}, and
Hierarchical cluster analysis using \code{hclust} (from stats package or flashClust if installed), and
render the tree with visNetwork, adding informations.
}
\examples{

\dontrun{

visHclust(iris, cutree = 3, nodesSize = 200)

visHclust(iris, cutree = 3,
  tooltipColumns = c(1, 5), nodesSize = 30,
  colorGroups = c("red", "blue", "green"))
  
  visHclust(iris, cutree = 8)\%>\% 
  visGroups(groupname = "cluster", color ="#00FF00", shape = "square")  \%>\% 
  visGroups(groupname = "individual", color ="#FF0000")
}

}
