% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visHclust.R
\name{visHclust}
\alias{visHclust}
\alias{visHclust.default}
\alias{visHclust.data.frame}
\alias{visHclust.dist}
\alias{visHclust.hclust}
\title{Visualize Hierarchical cluster analysis.}
\usage{
visHclust(object, ...)

\method{visHclust}{default}(object, ...)

\method{visHclust}{data.frame}(object, main = "", submain = "",
  footer = "", distColumns = NULL, distMethod = "euclidean",
  hclustMethod = "complete", cutree = 0, tooltipColumns = 1:ncol(object),
  colorEdges = "black", colorGroups = substr(rainbow(cutree), 1, 7),
  highlightNearest = TRUE, minNodeSize = 50, maxNodeSize = 200,
  nodesPopSize = TRUE, height = "600px", width = "100\%", export = TRUE,
  ...)

\method{visHclust}{dist}(object, data = NULL, main = "", submain = "",
  footer = "", cutree = 0, hclustMethod = "complete",
  tooltipColumns = NULL, colorEdges = "black",
  colorGroups = substr(rainbow(cutree), 1, 7), highlightNearest = TRUE,
  minNodeSize = 50, maxNodeSize = 200, nodesPopSize = TRUE,
  height = "600px", width = "100\%", export = TRUE, ...)

\method{visHclust}{hclust}(object, data = NULL, main = "", submain = "",
  footer = "", cutree = 0, tooltipColumns = NULL, colorEdges = "black",
  colorGroups = substr(rainbow(cutree), 1, 7), highlightNearest = TRUE,
  minNodeSize = 50, maxNodeSize = 200, nodesPopSize = TRUE,
  height = "600px", width = "100\%", export = TRUE, ...)
}
\arguments{
\item{object}{\code{hclust | dist | data.frame}.}

\item{...}{nothing}

\item{main}{For add a title. See \link{visNetwork}}

\item{submain}{For add a subtitle. See \link{visNetwork}}

\item{footer}{For add a footer. See \link{visNetwork}}

\item{distColumns}{\code{numeric} indice of columns used for compute distance. 
If NULL (default), keep all \code{numeric} and \code{integer} columns. 
If Not NULL, we keep only  \code{numeric} and \code{integer} columns}

\item{distMethod}{\code{character} the distance measure to be used for dist function. Default to 'euclidean'. See \code{\link[stats]{dist}}.}

\item{hclustMethod}{\code{character} the agglomeration method to be used for hclust function. Default to 'complete'. See \code{\link[stats]{hclust}}.}

\item{cutree}{\code{numeric} or \code{integer} desired number of groups. Default to 0}

\item{tooltipColumns}{\code{numeric} indice of columns used in tooltip. All by default.
So, we add a boxplot or a pie focus on sub-population and all population using \code{sparkline} package.}

\item{colorEdges}{\code{character} color of edges. Default to 'black'}

\item{colorGroups}{\code{character}, color for group in exa ("#00FF00"). Default rainbow.}

\item{highlightNearest}{\code{boolean} highlight sub-tree on click.}

\item{minNodeSize}{\code{numeric}, in case of \code{nodesPopSize}, minimum size of a node. Defaut to 15. Else, nodes size is minNodeSize + maxNodeSize / 2}

\item{maxNodeSize}{\code{numeric}, in case of \code{nodesPopSize}, maximum size of a node. Defaut to 30. Else, nodes size is minNodeSize + maxNodeSize / 2}

\item{nodesPopSize}{\code{boolean}, nodes sizes depends on population ? Default to FALSE}

\item{height}{\code{character}, default to "600px"}

\item{width}{\code{character}, default to "100\%"}

\item{export}{\code{boolean}, add button for export.}

\item{data}{\code{data.frame} data.frame with data.}
}
\description{
Visualize Hierarchical cluster analysis \code{hclust}. This function compute distance using \code{dist}, and
Hierarchical cluster analysis using \code{hclust} (from stats package or flashClust if installed), and
render the tree with visNetwork, adding informations.
}
\examples{

\dontrun{

visHclust(iris, cutree = 3)

visHclust(iris, cutree = 3,
  tooltipColumns = c(1, 5),
  colorGroups = c("red", "blue", "green"))
  
# no graphics on tooltip
visHclust(iris, cutree = 3, main = "Hclust on iris", 
  tooltipColumns = NULL)
  
  visHclust(iris, cutree = 8)\%>\% 
  visGroups(groupname = "group", color ="#00FF00", shape = "square")  \%>\% 
  visGroups(groupname = "individual", color ="#FF0000")
  
  #On dist
  visHclust(dist(iris[,1:4]), cutree = 3, data = iris)
  
  #On hclust
  visHclust(hclust(dist(iris[,1:4])), cutree = 3)
  
  
}

}
