% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclustTree.R
\name{visHclust}
\alias{visHclust}
\title{visNetwork for hclust}
\usage{
visHclust(data, colUseForDist = NULL, main = "", edgeColor = "black",
  nodeSize = 30, highlightNearest = TRUE, cutree = 0,
  detailsOnTooltips = TRUE, labelDraw = 1:ncol(data),
  colorGroup = substr(rainbow(cutree), 1, 7), hclustMethod = "complete",
  distMethode = "euclidean")
}
\arguments{
\item{data}{\code{data.frame} data.}

\item{colUseForDist}{\code{numeric} indice of columns to use for compute of dist.
No numeric columns are automaticaly excludes.}

\item{main}{\code{character} title.}

\item{edgeColor}{\code{character} color for edge.}

\item{nodeSize}{\code{numeric} size for nodes.}

\item{highlightNearest}{\code{boolean} highlight branch when you click.}

\item{cutree}{\code{numeric} number of group.}

\item{detailsOnTooltips}{\code{boolean} show details on tooltips (sparkline).}

\item{labelDraw}{\code{numeric} indice of columns to add in label, default all.}

\item{colorGroup}{\code{character}, color for group in exa ("#00FF00"). Default rainbow.}

\item{hclustMethod}{\code{character} method for hclust function. See \code{\link[stats]{hclust}} for help.}

\item{distMethode}{\code{character} method for dist function. See \code{\link[stats]{dist}} for help.}
}
\description{
visNetwork for hclust
}
\examples{

\dontrun{

visHclust(iris, cutree = 3, qualiSup = 5, nodeSize = 50)
visHclust(iris, cutree = 3,
  detailsOnTooltips = TRUE,
  qualiSup = 5, labelDraw = c(1, 5), nodeSize = 30,
  colorGroup = c("#DF0101", "#FF8000", "#D7DF01"))\%>\% 
  visGroups(groupname = "cluster", color ="#00FF00")  \%>\% 
  visGroups(groupname = "individual", color ="#FF0000")
}

}
