% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visMoveNode.R
\name{visMoveNode}
\alias{visMoveNode}
\title{Network visualization moveNode method}
\usage{
visMoveNode(graph, id, x, y)
}
\arguments{
\item{graph}{: a \code{\link[pkg:visNetwork]{visNetworkProxy}}  object}

\item{id}{: a node id}

\item{x}{: Number. x position, in canvas space}

\item{y}{: Number. y position, in canvas space}
}
\description{
For use moveNode() method in a shiny app. For full documentation, have a look at \link[pkg:visNetwork]{visDocumentation}.
}
\examples{
\dontrun{

# have a look to : 
shiny::runApp(system.file("shiny", package = "visNetwork"))

}

}
\references{
See online documentation \url{https://datastorm-open.github.io/visNetwork/}
}
\seealso{
\link[pkg:visNetwork]{visNodes} for nodes options, \link[pkg:visNetwork]{visEdges} for edges options, \link[pkg:visNetwork]{visGroups} for groups options, 
\link[pkg:visNetwork]{visLegend} for adding legend, \link[pkg:visNetwork]{visOptions} for custom option, \link[pkg:visNetwork]{visLayout} & \link[pkg:visNetwork]{visHierarchicalLayout} for layout, 
\link[pkg:visNetwork]{visPhysics} for control physics, \link[pkg:visNetwork]{visInteraction} for interaction, \link[pkg:visNetwork]{visNetworkProxy} & \link[pkg:visNetwork]{visFocus} & \link[pkg:visNetwork]{visFit} for animation within shiny,
\link[pkg:visNetwork]{visDocumentation}, \link[pkg:visNetwork]{visEvents}, \link[pkg:visNetwork]{visConfigure} ...
}
