% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visMoveNode.R
\name{visMoveNode}
\alias{visMoveNode}
\title{Network visualization moveNode method}
\usage{
visMoveNode(graph, id, x, y)
}
\arguments{
\item{graph}{: a \code{\link{visNetworkProxy}}  object}

\item{id}{: a node id}

\item{x}{: Number. x position, in canvas space}

\item{y}{: Number. y position, in canvas space}
}
\description{
For use moveNode() method in a shiny app. For full documentation, have a look at \link{visDocumentation}.
}
\examples{
\dontrun{

# have a look to : 
shiny::runApp(system.file("shiny", package = "visNetwork"))

}

}
\references{
See online documentation \url{http://datastorm-open.github.io/visNetwork/}
}
\seealso{
\link{visNodes} for nodes options, \link{visEdges} for edges options, \link{visGroups} for groups options, 
\link{visLegend} for adding legend, \link{visOptions} for custom option, \link{visLayout} & \link{visHierarchicalLayout} for layout, 
\link{visPhysics} for control physics, \link{visInteraction} for interaction, \link{visNetworkProxy} & \link{visFocus} & \link{visFit} for animation within shiny,
\link{visDocumentation}, \link{visEvents}, \link{visConfigure} ...
}
