% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visNearestNodes.R
\name{visNearestNodes}
\alias{visNearestNodes}
\title{Function to nearest nodes of a target node, with shiny only.}
\usage{
visNearestNodes(graph, target, maxpoints = 5, addDist = T)
}
\arguments{
\item{graph}{: a \code{ \link[visNetwork]{visNetworkProxy}}  object}

\item{target}{: name of shiny input returning target node id}

\item{maxpoints}{: Number of nearest nodes. Default to 5}

\item{addDist}{: If TRUE, add a column named dist_ that contains the distance from the coordinate to the point, in pixels.}
}
\description{
Function to nearest nodes of a target node, with shiny only.
}
\examples{
\dontrun{

# have a look to : 
shiny::runApp(system.file("shiny", package = "visNetwork"))

}

}
\references{
See online documentation \url{https://datastorm-open.github.io/visNetwork/}
}
\seealso{
\link[visNetwork]{visNodes} for nodes options,  \link[visNetwork]{visEdges} for edges options,  \link[visNetwork]{visGroups} for groups options, 
\link[visNetwork]{visLegend} for adding legend,  \link[visNetwork]{visOptions} for custom option,  \link[visNetwork]{visLayout} &  \link[visNetwork]{visHierarchicalLayout} for layout, 
\link[visNetwork]{visPhysics} for control physics,  \link[visNetwork]{visInteraction} for interaction,  \link[visNetwork]{visNetworkProxy} &  \link[visNetwork]{visFocus} &  \link[visNetwork]{visFit} for animation within shiny,
\link[visNetwork]{visDocumentation},  \link[visNetwork]{visEvents},  \link[visNetwork]{visConfigure} ...
}
