% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visNearestNodes.R
\name{visNearestNodes}
\alias{visNearestNodes}
\title{Function to nearest nodes of a target node, with shiny only.}
\usage{
visNearestNodes(graph, target, maxpoints = 5, addDist = T)
}
\arguments{
\item{graph}{: a \code{\link{visNetworkProxy}}  object}

\item{target}{: name of shiny input returning target node id}

\item{maxpoints}{: Number of nearest nodes. Default to 5}

\item{addDist}{: If TRUE, add a column named dist_ that contains the distance from the coordinate to the point, in pixels.}
}
\description{
Function to nearest nodes of a target node, with shiny only.
}
\examples{
\dontrun{

# have a look to : 
shiny::runApp(system.file("shiny", package = "visNetwork"))

}

}
\references{
See online documentation \url{http://datastorm-open.github.io/visNetwork/}
}
\seealso{
\link{visNodes} for nodes options, \link{visEdges} for edges options, \link{visGroups} for groups options, 
\link{visLegend} for adding legend, \link{visOptions} for custom option, \link{visLayout} & \link{visHierarchicalLayout} for layout, 
\link{visPhysics} for control physics, \link{visInteraction} for interaction, \link{visNetworkProxy} & \link{visFocus} & \link{visFit} for animation within shiny,
\link{visDocumentation}, \link{visEvents}, \link{visConfigure} ...
}
