% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/visNetwork.R
\name{visNetwork}
\alias{visNetwork}
\title{Network visualization}
\usage{
visNetwork(nodes, edges, highlight.nearest = TRUE, legend = FALSE,
  width = NULL, height = NULL)
}
\arguments{
\item{nodes}{: data.frame with nodes informations. Needed at least column "id".
See \url{http://visjs.org/docs/network.html#Nodes} and \url{http://visjs.org/docs/network.html#Nodes_configuration}
\itemize{
 \item{"id"}{ : id of the node, needed in edges information}
 \item{"label"}{ : label of the node}
 \item{"group"}{ : group of the node. Groups can be configure with \link{visGroups}}
 \item{"value"}{ : size of the node}
 \item{"title"}{ : tooltip of the node}
 \item{...}{}
}}

\item{edges}{: data.frame with edges informations. Needed at least columns "from" and "to".
See \url{http://visjs.org/docs/network.html#Edges} and \url{http://visjs.org/docs/network.html#Edges_configuration}
\itemize{
 \item{"from"}{ : node id of begin of the edge}
 \item{"to"}{ : node id of end of the edge}
 \item{"label"}{ : label of the edge}
 \item{"value"}{ : size of the node}
 \item{"title"}{ : tooltip of the node}
 \item{...}{}
}}

\item{highlight.nearest}{: Highlight nearest when clicking a node ? Default to TRUE. Based on \url{http://visjs.org/examples/network/29_neighbourhood_highlight.html}
This options use click event}

\item{legend}{: Boolean. Default to FALSE. A little bit experimental. Put a legend in case of groups.}
}
\description{
Network visualization using vis.js library
}
\examples{
# minimal example
nodes <- data.frame(id = 1:3)
edges <- data.frame(from = c(1,2), to = c(1,3))

visNetwork(nodes, edges, highlight.nearest = FALSE)

# more variables
nb = 15
nodes <- data.frame(id = 1:nb, label = paste("Label", 1:nb),
 group = sample(1:nb, nb, replace = TRUE), value = 1:nb,
 title = paste0("<p>", 1:nb,"<br>Tooltip !</p>"), stringsAsFactors = FALSE)

edges <- data.frame(from = trunc(runif(nb)*(nb-1))+1,
 to = trunc(runif(nb)*(nb-1))+1,
 value = rnorm(nb, 10), label = paste("Edge", 1:nb),
 title = paste0("<p>", 1:nb,"<br>Edge Tooltip !</p>"))

# simple network
visNetwork(nodes, edges)

# try a legend...
visNetwork(nodes, edges, legend = TRUE)

# directed network
visNetwork(nodes, edges) \%>\% visEdges(style = "arrow")

# custom navigation
visNetwork(nodes, edges) \%>\% visOptions(navigation = TRUE)

# data Manipulation
visNetwork(nodes, edges) \%>\% visOptions(dataManipulation = TRUE)

# Hierarchical Layout
visNetwork(nodes, edges) \%>\% visHierarchicalLayout()

# freeze network
visNetwork(nodes, edges) \%>\% visOptions(dragNetwork = FALSE, dragNodes = FALSE)

# clustering
nodes <- data.frame(id = 1:100)
edges <- data.frame(from = round(runif(120)*100), to = round(runif(120)*100))

visNetwork(nodes, edges) \%>\%
 visClustering(initialMaxNodes = 50, nodeScaling = list(width = 50, height = 50, radius = 50))
}
\seealso{
\link{visOptions}, \link{visNodes}, \link{visEdges}, \link{visGroups}, \link{visEvents}, ...
}

