% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visOptions.R
\name{visOptions}
\alias{visOptions}
\title{Network visualization general options}
\usage{
visOptions(graph, width = NULL, height = NULL, highlightNearest = FALSE,
  nodesIdSelection = FALSE, selectedBy = NULL, autoResize = NULL,
  clickToUse = NULL, manipulation = NULL)
}
\arguments{
\item{graph}{: a visNetwork object}

\item{width}{: String. Default to "100\%". The width of the network in pixels or as a percentage.}

\item{height}{: String. Default to "100\%". The height of the network in pixels or as a percentage.}

\item{highlightNearest}{: Custom Option. Just a Boolean, or a named list. Default to false. Highlight nearest when clicking a node ? This options use click event. Not available for DOT and Gephi.
\itemize{
 \item{"enabled"}{ : Boolean. Default to false. Activated or not ?.}
 \item{"degree"}{ : Integer. Degree of depth of nodes to be colored. Default to 1}
}}

\item{nodesIdSelection}{:  Custom Option. Just a Boolean, or a named list. Default to false. Add an id node selection creating an HTML select element. This options use click event. Not available for DOT and Gephi.
\itemize{
 \item{"enabled"}{ : Boolean. Default to false. Activated or not ?.}
 \item{"values}{ : Optional. Vector of possible values. Defaut to all id in nodes data.frame.}
 \item{"selected"}{ : Optional. Integer/Character. Initial id selection. Defaut to NULL}
 \item{"style"}{ : Character. HTML style of list. Default to 'width: 150px; height: 26px'. Optional.}
}}

\item{selectedBy}{: Custom option. Character or a named list. Add a multiple selection based on column of node data.frame creating an HTML select element. Not available for DOT and Gephi.
\itemize{
 \item{"variable"}{ : Character. Column name of selection variable.}
 \item{"values}{ : Optional. Vector of possible values. Defaut to all values in nodes data.frame.}
 \item{"selected"}{ : Optional. Integer/Character. Initial selection. Defaut to NULL}
 \item{"style"}{ : Optional. Character. HTML style of list. Default to 'width: 150px; height: 26px'. Optional.}
}}

\item{autoResize}{: Boolean. Default to true. If true, the Network will automatically detect when its container is resized, and redraw itself accordingly. If false, the Network can be forced to repaint after its container has been resized using the function redraw() and setSize().}

\item{clickToUse}{: Boolean. Default to false. When a Network is configured to be clickToUse, it will react to mouse, touch, and keyboard events only when active. When active, a blue shadow border is displayed around the Network. The Network is set active by clicking on it, and is changed to inactive again by clicking outside the Network or by pressing the ESC key.}

\item{manipulation}{: Just a Boolean}
}
\description{
Network visualization general options. For full documentation, have a look at \link{visDocumentation}.
}
\examples{
nodes <- data.frame(id = 1:15, label = paste("Label", 1:15),
 group = sample(LETTERS[1:3], 15, replace = TRUE))

edges <- data.frame(from = trunc(runif(15)*(15-1))+1,
 to = trunc(runif(15)*(15-1))+1)

###################
# highlight nearest
###################

visNetwork(nodes, edges) \%>\% visOptions(highlightNearest = TRUE)
visNetwork(nodes, edges) \%>\% visOptions(highlightNearest = list(enabled = TRUE, degree = 2))

##########################
# nodesIdSelection
##########################

visNetwork(nodes, edges) \%>\%
 visOptions(highlightNearest = TRUE, nodesIdSelection = TRUE)

# add a default selected node ?
visNetwork(nodes, edges) \%>\%
 visOptions(highlightNearest = TRUE,
 nodesIdSelection = list(enabled = TRUE, selected = "1"))

# subset on id values ?
visNetwork(nodes, edges) \%>\%
 visOptions(highlightNearest = TRUE,
 nodesIdSelection = list(enabled = TRUE,
   selected = "2",
   values = c(2:10)))

# some style
visNetwork(nodes, edges) \%>\%
 visOptions(highlightNearest = TRUE,
 nodesIdSelection = list(enabled = TRUE, style = 'width: 200px; height: 26px;
   background: #f8f8f8;
   color: darkblue;
   border:none;
   outline:none;'))

##########################
# selectedBy
##########################

visNetwork(nodes, edges) \%>\%
 visOptions(selectedBy = "group")

# add a default value ?
visNetwork(nodes, edges) \%>\%
 visOptions(selectedBy = list(variable = "group", selected = "A"))

# subset on values ?
visNetwork(nodes, edges) \%>\%
 visOptions(selectedBy = list(variable = "group",
   selected = "C",
   values = c("A", "C")))

# add some style
visNetwork(nodes, edges) \%>\%
 visOptions(selectedBy = list(variable = "group", style = 'width: 200px; height: 26px;
   background: #f8f8f8;
   color: darkblue;
   border:none;
   outline:none;'))

# can also be on new column
nodes$sample <- sample(c("sample 1", "sample 2"), nrow(nodes), replace = TRUE)
visNetwork(nodes, edges) \%>\%
 visOptions(selectedBy = "sample")
}
\seealso{
\link{visNodes} for nodes options, \link{visEdges} for edges options, \link{visGroups} for groups options,
\link{visLegend} for adding legend, \link{visOptions} for custom option, \link{visLayout} & \link{visHierarchicalLayout} for layout,
\link{visPhysics} for control physics, \link{visInteraction} for interaction, \link{visDocumentation}, \link{visEvents}, \link{visConfigure} ...
}

