% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/visRemoveEdges.R
\name{visRemoveEdges}
\alias{visRemoveEdges}
\title{Function to remove edges from network, with shiny only.}
\usage{
visRemoveEdges(graph, id)
}
\arguments{
\item{graph}{: a \code{\link{visNetworkProxy}}  object}

\item{id}{: vector of id, edges to remove}
}
\description{
Function to remove edges from network, with shiny only.
}
\examples{
\dontrun{

# have a look to :
shiny::runApp(system.file("shiny", package = "visNetwork"))

}
}
\seealso{
\link{visNodes} for nodes options, \link{visEdges} for edges options, \link{visGroups} for groups options,
\link{visLegend} for adding legend, \link{visOptions} for custom option, \link{visLayout} & \link{visHierarchicalLayout} for layout,
\link{visPhysics} for control physics, \link{visInteraction} for interaction, \link{visNetworkProxy} & \link{visFocus} & \link{visFit} for animation within shiny,
\link{visDocumentation}, \link{visEvents}, \link{visConfigure} ...
}

