% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visRemoveNodes.R
\name{visRemoveNodes}
\alias{visRemoveNodes}
\title{Function to remove nodes from network, with shiny only.}
\usage{
visRemoveNodes(graph, id, updateOptions = TRUE, legend = FALSE)
}
\arguments{
\item{graph}{: a \code{\link{visNetworkProxy}}  object}

\item{id}{: vector of id, nodes to remove}

\item{updateOptions}{: Boolean. Update options (nodesIdSelection & selectedBy) if needed ? Default to TRUE.}

\item{legend}{: Boolean. Remove nodes on legend ? Default to FALSE}
}
\description{
Function to remove nodes from network, with shiny only.
}
\examples{
\dontrun{

# have a look to : 
shiny::runApp(system.file("shiny", package = "visNetwork"))

}

}
\references{
See online documentation \url{http://datastorm-open.github.io/visNetwork/}
}
\seealso{
\link{visNodes} for nodes options, \link{visEdges} for edges options, \link{visGroups} for groups options, 
\link{visLegend} for adding legend, \link{visOptions} for custom option, \link{visLayout} & \link{visHierarchicalLayout} for layout, 
\link{visPhysics} for control physics, \link{visInteraction} for interaction, \link{visNetworkProxy} & \link{visFocus} & \link{visFit} for animation within shiny,
\link{visDocumentation}, \link{visEvents}, \link{visConfigure} ...
}
