% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visSelectNodes.R
\name{visSelectNodes}
\alias{visSelectNodes}
\title{Function to select node(s) from network, with shiny only.}
\usage{
visSelectNodes(graph, id, highlightEdges = TRUE, clickEvent = TRUE)
}
\arguments{
\item{graph}{: a \code{ \link[visNetwork]{visNetworkProxy}}  object}

\item{id}{: vector of id, node(s) to select}

\item{highlightEdges}{: Boolean. highlight Edges also ? Default to TRUE}

\item{clickEvent}{: Boolean. Launch click event ? (highlightNearest for example) Default to TRUE}
}
\description{
Function to select node(s) from network, with shiny only.
}
\examples{
\dontrun{

# have a look to : 
shiny::runApp(system.file("shiny", package = "visNetwork"))

}

}
\references{
See online documentation \url{https://datastorm-open.github.io/visNetwork/}
}
\seealso{
\link[visNetwork]{visNodes} for nodes options,  \link[visNetwork]{visEdges} for edges options,  \link[visNetwork]{visGroups} for groups options, 
\link[visNetwork]{visLegend} for adding legend,  \link[visNetwork]{visOptions} for custom option,  \link[visNetwork]{visLayout} &  \link[visNetwork]{visHierarchicalLayout} for layout, 
\link[visNetwork]{visPhysics} for control physics,  \link[visNetwork]{visInteraction} for interaction,  \link[visNetwork]{visNetworkProxy} &  \link[visNetwork]{visFocus} &  \link[visNetwork]{visFit} for animation within shiny,
\link[visNetwork]{visDocumentation},  \link[visNetwork]{visEvents},  \link[visNetwork]{visConfigure} ...
}
