% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visSetData.R
\name{visSetData}
\alias{visSetData}
\title{Network visualization setData method}
\usage{
visSetData(graph, nodes = NULL, edges = NULL)
}
\arguments{
\item{graph}{: a \code{ \link[visNetwork]{visNetworkProxy}}  object}

\item{nodes}{: data.frame with nodes informations. Needed at least column "id". See  \link[visNetwork]{visNodes}}

\item{edges}{: data.frame with edges informations. Needed at least columns "from" and "to". See  \link[visNetwork]{visEdges}}
}
\description{
For use setData() method in a shiny app. For full documentation, have a look at  \link[visNetwork]{visDocumentation}.
}
\examples{
\dontrun{

# have a look to : 
shiny::runApp(system.file("shiny", package = "visNetwork"))

}

}
\references{
See online documentation \url{https://datastorm-open.github.io/visNetwork/}
}
\seealso{
\link[visNetwork]{visNodes} for nodes options,  \link[visNetwork]{visEdges} for edges options,  \link[visNetwork]{visGroups} for groups options, 
\link[visNetwork]{visLegend} for adding legend,  \link[visNetwork]{visOptions} for custom option,  \link[visNetwork]{visLayout} &  \link[visNetwork]{visHierarchicalLayout} for layout, 
\link[visNetwork]{visPhysics} for control physics,  \link[visNetwork]{visInteraction} for interaction,  \link[visNetwork]{visNetworkProxy} &  \link[visNetwork]{visFocus} &  \link[visNetwork]{visFit} for animation within shiny,
\link[visNetwork]{visDocumentation},  \link[visNetwork]{visEvents},  \link[visNetwork]{visConfigure} ...
}
