% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visSetData.R
\name{visSetData}
\alias{visSetData}
\title{Network visualization setData method}
\usage{
visSetData(graph, nodes = NULL, edges = NULL)
}
\arguments{
\item{graph}{: a \code{\link{visNetworkProxy}}  object}

\item{nodes}{: data.frame with nodes informations. Needed at least column "id". See \link{visNodes}}

\item{edges}{: data.frame with edges informations. Needed at least columns "from" and "to". See \link{visEdges}}
}
\description{
For use setData() method in a shiny app. For full documentation, have a look at \link{visDocumentation}.
}
\examples{
\dontrun{

# have a look to : 
shiny::runApp(system.file("shiny", package = "visNetwork"))

}

}
\references{
See online documentation \url{http://datastorm-open.github.io/visNetwork/}
}
\seealso{
\link{visNodes} for nodes options, \link{visEdges} for edges options, \link{visGroups} for groups options, 
\link{visLegend} for adding legend, \link{visOptions} for custom option, \link{visLayout} & \link{visHierarchicalLayout} for layout, 
\link{visPhysics} for control physics, \link{visInteraction} for interaction, \link{visNetworkProxy} & \link{visFocus} & \link{visFit} for animation within shiny,
\link{visDocumentation}, \link{visEvents}, \link{visConfigure} ...
}
