% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visSetSelection.R
\name{visSetSelection}
\alias{visSetSelection}
\title{Function to select edge(s) / node(s) from network, with shiny only.}
\usage{
visSetSelection(
  graph,
  nodesId = NULL,
  edgesId = NULL,
  unselectAll = TRUE,
  highlightEdges = TRUE,
  clickEvent = TRUE
)
}
\arguments{
\item{graph}{: a \code{\link{visNetworkProxy}}  object}

\item{nodesId}{: vector of id, nodes(s) to select}

\item{edgesId}{: vector of id, edges(s) to select}

\item{unselectAll}{: Boolean. Unselect all nodes & edges before current selection ? Default to TRUE}

\item{highlightEdges}{: Boolean. highlight Edges also ? Default to TRUE}

\item{clickEvent}{: Boolean. Launch click event ? (highlightNearest for example) Default to TRUE}
}
\description{
Function to select edge(s) / node(s) from network, with shiny only.
}
\examples{
\dontrun{

# have a look to : 
shiny::runApp(system.file("shiny", package = "visNetwork"))

}

}
\references{
See online documentation \url{http://datastorm-open.github.io/visNetwork/}
}
\seealso{
\link{visNodes} for nodes options, \link{visEdges} for edges options, \link{visGroups} for groups options, 
\link{visLegend} for adding legend, \link{visOptions} for custom option, \link{visLayout} & \link{visHierarchicalLayout} for layout, 
\link{visPhysics} for control physics, \link{visInteraction} for interaction, \link{visNetworkProxy} & \link{visFocus} & \link{visFit} for animation within shiny,
\link{visDocumentation}, \link{visEvents}, \link{visConfigure} ...
}
