% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visSetTitle.R
\name{visSetTitle}
\alias{visSetTitle}
\title{Set title, subtitle, and footer using \code{visNetworkProxy}}
\usage{
visSetTitle(graph, main = NULL, submain = NULL, footer = NULL)
}
\arguments{
\item{graph}{: a \code{\link{visNetworkProxy}}  object}

\item{main}{: For add a title. Character or a named list.
\itemize{
 \item{"text"}{ : Character. Title.}
 \item{"style"}{ : Optional. Character. HTML style of title.}
 \item{'hidden'}{ : Optional. Boolean. Force title to be hidden} 
}}

\item{submain}{: For add a subtitle. Character or a named list.
\itemize{
 \item{"text"}{ : Character. Subtitle.}
 \item{"style"}{ : Optional. Character. HTML style of submain.}
 \item{'hidden'}{ : Optional. Boolean. Force submain to be hidden} 
}}

\item{footer}{: For add a footer. Character or a named list.
\itemize{
 \item{"text"}{ : Character. footer.}
 \item{"style"}{ : Optional. Character. HTML style of footer.}
 \item{'hidden'}{ : Optional. Boolean. Force footer to be be hidden} 
}}
}
\description{
Set title, subtitle, and footer using \code{visNetworkProxy}
}
\examples{
\dontrun{

# have a look to : 
shiny::runApp(system.file("shiny", package = "visNetwork"))

}

}
\references{
See online documentation \url{http://datastorm-open.github.io/visNetwork/}
}
\seealso{
\link{visNodes} for nodes options, \link{visEdges} for edges options, \link{visGroups} for groups options, 
\link{visLegend} for adding legend, \link{visOptions} for custom option, \link{visLayout} & \link{visHierarchicalLayout} for layout, 
\link{visPhysics} for control physics, \link{visInteraction} for interaction, \link{visNetworkProxy} & \link{visFocus} & \link{visFit} for animation within shiny,
\link{visDocumentation}, \link{visEvents}, \link{visConfigure} ...
}
