% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visTree.R
\name{visTree}
\alias{visTree}
\title{Plotting rpart objects}
\usage{
visTree(tree, tooltipDelay = 500, rules = TRUE, main = "",
  highlightNearest = TRUE, legend = TRUE, legendWidth = 0.2,
  legendNcol = 3, collapse = TRUE, colorVar = NULL,
  colorModality = NULL, colorEdge = "#8181F7", direction = "UD",
  digits = 3)
}
\arguments{
\item{tree}{\code{rpart}, rpart object to plot}

\item{tooltipDelay}{\code{numeric}, delay before tooltips 
apparition in millisecond default 500}

\item{rules}{\code{boolean}, add rules in tooltips default TRUE}

\item{main}{\code{character}, Graph title}

\item{highlightNearest}{\code{boolean}, Highlight nearest default TRUE}

\item{legend}{\code{boolean}, add legend default TRUE}

\item{legendWidth}{\code{numeric}, legend width between 0 and 1 default 0.2}

\item{legendNcol}{\code{numeric}, number of column for legend default 3}

\item{collapse}{\code{boolean}, Said if graph is collapse or not, try to
Double click on a node default TRUE}

\item{colorVar}{\code{data.frame} 2 columns :
variable contain names of variables X
color contain colors in hexa}

\item{colorModality}{\code{data.frame} 2 columns :
modality contain levels of Y
color contain colors in hexa}

\item{colorEdge}{\code{character} color of edge in hexa default #8181F7}

\item{direction}{\code{character}, The direction of the hierarchical layout.
The available options are: UD, DU, LR, RL. To simplify:
up-down, down-up, left-right, right-left. Default UD}

\item{digits}{\code{numeric}, number of digits

#@examples


library(rpart)
res<- rpart::rpart(Species~.,data=iris)
#Basic classification tree
visTree(res)
visTree(res, direction = 'LR')
visTree(res, legendWidth = 0.1, legendNcol = 1)
visTree(res, main = "Title")
#Basic regression tree
res<- rpart::rpart(Petal.Length~.,data=iris)
visTree(res)
visTree(res, rules = FALSE, tooltipDelay = 0)

#Complexe tree
data("solder")
res<- rpart::rpart(Opening~.,data=solder, control = (rpart.control(cp = 0.000005)))
visTree(res, colorEdge = "#000099")

#Color nodes
colorVar <- data.frame(variable = names(solder), color = c("#339933", "#b30000","#4747d1",
                                                           "#88cc00", "#9900ff","#247856"))

colorModality <- data.frame(modality = unique(solder$Opening), color = 
                              c("#AA00AA", "#CDAD15", "#213478"))

visTree(res, colorEdge = "#000099", colorVar = colorVar, 
        colorModality = colorModality)}
}
\description{
Plotting rpart objects
}

