% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visTree.R
\name{visTreeEditor}
\alias{visTreeEditor}
\title{Run and edit a visTree, and get back in R}
\usage{
visTreeEditor(data, ...)
}
\arguments{
\item{data}{\code{rpart or data.drame}}

\item{...}{all arguments except \code{object} present in \link{visTreeModuleServer}}
}
\description{
Needed packages : shiny, rpart, colourpicker, shinyWidgets
}
\examples{

\dontrun{

net <- visTreeEditor(data = iris)
net <- visTreeEditor(data = rpart(iris), main = "visTree Editor")
net <- visTreeEditor(data = rpart(iris), tooltip_data = iris, 
    main = "visTree Editor")
net

}

}
\references{
See online documentation \url{http://datastorm-open.github.io/visNetwork/}
}
\seealso{
\link{visTree}, \link{visTreeModuleServer}, \link{visNetworkEditor}
}
